requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
PyNaCl==1.5.0
async-timeout==3.0.1
youtube-dl==2021.12.17
ffmpeg==1.4
python-dotenv==0.19.2
pytube==12.0.0
youtube-search==2.1.0